+++ examples/credentials_user_switch_get.py
@@ -74,7 +74,7 @@ from nd_python.parsers.parser_nd_domain import parser_nd_domain
 from nd_python.parsers.parser_nd_ip4 import parser_nd_ip4
 from nd_python.parsers.parser_nd_password import parser_nd_password
 from nd_python.parsers.parser_nd_username import parser_nd_username
-from nd_python.validators.credentials_user_switch_get import CredentialsUserSwitchGetConfigValidator
+from nd_python.validators.credentials.user_switch_get import CredentialsUserSwitchGetConfigValidator
 from pydantic import ValidationError
