+++ lib/nd_python/validators/credentials/robot_switch_save.py
