+++ lib/nd_python/validators/credentials/user_switch_get.py
